python-server/requirements.txt
pydantic==2.3.0
python-socketio==5.9.0
uvicorn==0.23.2
python-dotenv==1.0.0
fastapi==0.104.1